Predict_model/requirements.txt
flask
pillow
tensorflow
keras
pytest

